feat: 完善数据导入 用户权限

+++ app/models/point_record.py
@@ -3,9 +3,9 @@ from decimal import Decimal
 from typing import Optional
 from uuid import UUID
 
-from sqlalchemy import ForeignKey, Integer, String, DateTime, Numeric, func
+from sqlalchemy import ForeignKey, Integer, String, DateTime, Numeric, func, BigInteger
 from sqlalchemy.dialects.postgresql import UUID as PG_UUID
-from sqlalchemy.orm import Mapped, mapped_column, DeclarativeBase
+from sqlalchemy.orm import Mapped, mapped_column
 
 from app.models.user import Base
 
@@ -13,7 +13,7 @@ from app.models.user import Base
 class PointRecordModel(Base):
     __tablename__ = "point_records"
     __table_args__ = {"schema": "interview"}
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
     user_id: Mapped[UUID] = mapped_column(
         PG_UUID(as_uuid=True),
         ForeignKey("interview.users.user_id", ondelete="CASCADE"),
